+++ test/aliases_test.dart
@@ -1,0 +1,20 @@
+// SPDX-FileCopyrightText: (c) 2021 Art Galkin <github.com/rtmigo>
+// SPDX-License-Identifier: BSD-3-Clause
+
+
+import 'package:test/test.dart';
+import 'package:xrandom/xrandom.dart';
+
+void main() {
+
+  test('Xrandom', () {
+
+    expect(Xrandom() is Xorshift32, true);
+    expect(Xrandom(1) is Xorshift32, true);
+    expect(Xrandom.expected() is Xorshift32, true);
+
+  });
+
+
+
+}