+++ CHANGELOG.md
@@ -1,16 +1,22 @@
 # 0.1.0
 
+- Added Xrandom and Xrandom64 aliases
+
 - Added xoshiro128++ 
+
+- deterministic() replaced to expected()
 
 - Speed improvements
 
-- Changed the order of nextBool() results
+- Changed the order in which the results appear nextBool() results appear
 
 - Improved the accuracy of nextDouble() in Xorshift64 and 
   Xorshift120p for exact matching of reference values
     
 - Changed the order of nextInt32() results in Xorshift64 and 
   Xorshift120p
+ 
+  
 
 # 0.0.2